+++ packages/app/src/systems/DApp/hooks/useTxApprove.tsx
@@ -1,10 +1,12 @@
 import { useInterpret, useSelector } from '@xstate/react';
 import type { TransactionRequest } from 'fuels';
+import { bn } from 'fuels';
 
 import type { TxApproveMachineState } from '../machines';
 import { txApproveMachine } from '../machines';
 
 import { useAccount } from '~/systems/Account';
+import { getCoinOutputsFromTx, getGroupedErrors } from '~/systems/Transaction';
 
 const selectors = {
   isUnlocking: (state: TxApproveMachineState) => state.matches('unlocking'),
@@ -26,6 +28,19 @@ export function useTxApprove() {
   const tx = useSelector(service, selectors.tx);
   const receipts = useSelector(service, selectors.receipts);
   const txDryRunError = useSelector(service, selectors.txDryRunError);
+
+  const coinOutputs = getCoinOutputsFromTx(tx);
+  const outputsToSend = coinOutputs.filter(
+    (value) => value.to !== account?.publicKey
+  );
+  const outputAmount = outputsToSend.reduce(
+    (acc, value) => acc.add(value.amount),
+    bn(0)
+  );
+  const groupedErrors = getGroupedErrors(
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    (txDryRunError as any)?.response?.errors
+  );
 
   function startApprove() {
     send('START_APPROVE');
@@ -57,5 +72,9 @@ export function useTxApprove() {
     tx,
     receipts,
     txDryRunError,
+    coinOutputs,
+    outputsToSend,
+    outputAmount,
+    groupedErrors,
   };
 }
